requirements.txt
# This requirements.txt is auto-generated from Pipfile for deployment compatibility
# For development, use: pipenv install
# To update this file, run: pipenv requirements > requirements.txt

docker
gradio[mcp]
textblob
langchain-chroma>=0.1.2
langchain-openai
langchain-text-splitters 
langchain-community 
langchain-core
openai
tiktoken
chromadb>=0.4.1
unstructured[md]    
numpy<2.0
pandas
python-dotenv
mcp[cli]
fastmcp
duckduckgo-search
tavily-python
requests
beautifulsoup4
aiohttp
fastapi
uvicorn
click
python-multipart
starlette